Move all services to services folder
This fixes #35
