+++ src/index.ts
@@ -1,8 +1,11 @@
-export { ConfigurationService } from './configuration.service';
-export { APIService } from './api.service';
-export { AuthenticationService } from './authentication.service';
-export { ResourceService } from './resource.service';
-export { CommentsService } from './comments.service';
+export * from './interfaces';
+export { APIService } from './services/api.service';
+export { AuthenticationService } from './services/authentication.service';
+export { CacheService } from './services/cache.service';
+export { ConfigurationService } from './services/configuration.service';
+export { CommentsService } from './services/comments.service';
+export { LoadingService } from './services/loading.service';
+export { ResourceService } from './services/resource.service';
 export { Services } from './services';
 export { RESTAPIModule } from './module';
 export { PloneViews, InterfaceMarker, TypeMarker } from './traversal';
@@ -18,4 +21,3 @@ export { Comments, Comment, CommentAdd } from './components/comments';
 export { GlobalNavigation } from './components/global.navigation';
 export { Navigation } from './components/navigation';
 export { NavigationLevel } from './components/navigation.level';
-export * from './interfaces';
