+++ src/services/configuration.service.ts
