+++ src/views/login.spec.ts
@@ -7,19 +7,19 @@ import { FormsModule } from '@angular/forms';
 
 import { APP_BASE_HREF } from '@angular/common';
 
-import { ConfigurationService } from '../configuration.service';
-import { APIService } from '../api.service';
-import { AuthenticationService } from '../authentication.service';
-import { ResourceService } from '../resource.service';
-import { CommentsService } from '../comments.service';
-import { NavigationService } from '../navigation.service';
-import { CacheService } from '../cache.service';
+import { ConfigurationService } from '../services/configuration.service';
+import { APIService } from '../services/api.service';
+import { AuthenticationService } from '../services/authentication.service';
+import { CacheService } from '../services/cache.service';
+import { CommentsService } from '../services/comments.service';
+import { LoadingService } from '../services/loading.service';
+import { NavigationService } from '../services/navigation.service';
+import { ResourceService } from '../services/resource.service';
 import { Services } from '../services';
 import { Traverser, TraversalModule, Resolver, Marker, Normalizer } from 'angular-traversal';
 import { TypeMarker, RESTAPIResolver, PloneViews, FullPathNormalizer } from '../traversal';
 import { LoginView } from './login';
 import { ViewView } from './view';
-import { LoadingService } from '../loading.service';
 
 describe('LoginView', () => {
   let component: LoginView;
