+++ src/views/view.spec.ts
@@ -5,18 +5,18 @@ import {
 } from '@angular/common/http/testing';
 import { APP_BASE_HREF } from '@angular/common';
 
-import { ConfigurationService } from '../configuration.service';
-import { APIService } from '../api.service';
-import { AuthenticationService } from '../authentication.service';
-import { ResourceService } from '../resource.service';
-import { CommentsService } from '../comments.service';
-import { NavigationService } from '../navigation.service';
+import { ConfigurationService } from '../services/configuration.service';
+import { APIService } from '../services/api.service';
+import { AuthenticationService } from '../services/authentication.service';
+import { ResourceService } from '../services/resource.service';
+import { CommentsService } from '../services/comments.service';
+import { NavigationService } from '../services/navigation.service';
 import { Services } from '../services';
 import { Traverser, TraversalModule, Resolver, Marker, Normalizer } from 'angular-traversal';
 import { TypeMarker, RESTAPIResolver, PloneViews, FullPathNormalizer } from '../traversal';
 import { ViewView } from './view';
-import { CacheService } from '../cache.service';
-import { LoadingService } from '../loading.service';
+import { CacheService } from '../services/cache.service';
+import { LoadingService } from '../services/loading.service';
 
 
 describe('ViewView', () => {
